requirements.txt
fastapi[all]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv